+++ google/cloud/google_cloud_cpp_common.bzl
@@ -52,7 +52,6 @@ google_cloud_cpp_common_hdrs = [
     "internal/random.h",
     "internal/retry_policy.h",
     "internal/setenv.h",
-    "internal/stream_range.h",
     "internal/strerror.h",
     "internal/throw_delegate.h",
     "internal/tuple.h",
@@ -65,6 +64,7 @@ google_cloud_cpp_common_hdrs = [
     "polling_policy.h",
     "status.h",
     "status_or.h",
+    "stream_range.h",
     "terminate_handler.h",
     "tracing_options.h",
     "version.h",
