+++ google/cloud/stream_range_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/internal/stream_range.h"
+#include "google/cloud/stream_range.h"
 #include "google/cloud/testing_util/status_matchers.h"
 #include "absl/memory/memory.h"
 #include <gmock/gmock.h>
@@ -22,7 +22,6 @@
 namespace google {
 namespace cloud {
 inline namespace GOOGLE_CLOUD_CPP_NS {
-namespace internal {
 namespace {
 
 using ::google::cloud::testing_util::StatusIs;
@@ -39,14 +38,14 @@ TEST(StreamRange, DefaultConstructed) {
 
 TEST(StreamRange, MoveOnly) {
   auto const reader = [] { return Status{}; };
-  StreamRange<int> sr = MakeStreamRange<int>(reader);
+  StreamRange<int> sr = internal::MakeStreamRange<int>(reader);
   StreamRange<int> move_construct = std::move(sr);
-  StreamRange<int> move_assign = MakeStreamRange<int>(reader);
+  StreamRange<int> move_assign = internal::MakeStreamRange<int>(reader);
   move_assign = std::move(move_construct);
 }
 
 TEST(StreamRange, EmptyRange) {
-  StreamRange<int> sr = MakeStreamRange<int>([] { return Status{}; });
+  StreamRange<int> sr = internal::MakeStreamRange<int>([] { return Status{}; });
   auto it = sr.begin();
   auto end = sr.end();
   EXPECT_EQ(it, end);
@@ -56,12 +55,12 @@ TEST(StreamRange, EmptyRange) {
 
 TEST(StreamRange, OneElement) {
   auto counter = 0;
-  auto reader = [&counter]() -> StreamReader<int>::result_type {
+  auto reader = [&counter]() -> internal::StreamReader<int>::result_type {
     if (counter++ < 1) return 42;
     return Status{};
   };
 
-  StreamRange<int> sr = MakeStreamRange<int>(std::move(reader));
+  StreamRange<int> sr = internal::MakeStreamRange<int>(std::move(reader));
   auto it = sr.begin();
   EXPECT_NE(it, sr.end());
   EXPECT_TRUE(*it);
@@ -71,11 +70,11 @@ TEST(StreamRange, OneElement) {
 }
 
 TEST(StreamRange, OneError) {
-  auto reader = []() -> StreamReader<int>::result_type {
+  auto reader = []() -> internal::StreamReader<int>::result_type {
     return Status(StatusCode::kUnknown, "oops");
   };
 
-  StreamRange<int> sr = MakeStreamRange<int>(std::move(reader));
+  StreamRange<int> sr = internal::MakeStreamRange<int>(std::move(reader));
   auto it = sr.begin();
   EXPECT_NE(it, sr.end());
   EXPECT_FALSE(*it);
@@ -86,12 +85,12 @@ TEST(StreamRange, OneError) {
 
 TEST(StreamRange, FiveElements) {
   auto counter = 0;
-  auto reader = [&counter]() -> StreamReader<int>::result_type {
+  auto reader = [&counter]() -> internal::StreamReader<int>::result_type {
     if (counter++ < 5) return counter;
     return Status{};
   };
 
-  StreamRange<int> sr = MakeStreamRange<int>(std::move(reader));
+  StreamRange<int> sr = internal::MakeStreamRange<int>(std::move(reader));
   std::vector<int> v;
   for (StatusOr<int>& x : sr) {
     EXPECT_TRUE(x);
@@ -102,12 +101,12 @@ TEST(StreamRange, FiveElements) {
 
 TEST(StreamRange, PostFixIteration) {
   auto counter = 0;
-  auto reader = [&counter]() -> StreamReader<int>::result_type {
+  auto reader = [&counter]() -> internal::StreamReader<int>::result_type {
     if (counter++ < 5) return counter;
     return Status{};
   };
 
-  StreamRange<int> sr = MakeStreamRange<int>(std::move(reader));
+  StreamRange<int> sr = internal::MakeStreamRange<int>(std::move(reader));
   std::vector<int> v;
   // NOLINTNEXTLINE(modernize-loop-convert)
   for (auto it = sr.begin(); it != sr.end(); it++) {
@@ -119,13 +118,13 @@ TEST(StreamRange, PostFixIteration) {
 
 TEST(StreamRange, Distance) {
   // Empty range
-  StreamRange<int> sr = MakeStreamRange<int>([] { return Status{}; });
+  StreamRange<int> sr = internal::MakeStreamRange<int>([] { return Status{}; });
   EXPECT_EQ(0, std::distance(sr.begin(), sr.end()));
 
   // Range of one element
   auto counter = 0;
-  StreamRange<int> one =
-      MakeStreamRange<int>([&counter]() -> StreamReader<int>::result_type {
+  StreamRange<int> one = internal::MakeStreamRange<int>(
+      [&counter]() -> internal::StreamReader<int>::result_type {
         if (counter++ < 1) return counter;
         return Status{};
       });
@@ -133,8 +132,8 @@ TEST(StreamRange, Distance) {
 
   // Range of five elements
   counter = 0;
-  StreamRange<int> five =
-      MakeStreamRange<int>([&counter]() -> StreamReader<int>::result_type {
+  StreamRange<int> five = internal::MakeStreamRange<int>(
+      [&counter]() -> internal::StreamReader<int>::result_type {
         if (counter++ < 5) return counter;
         return Status{};
       });
@@ -143,12 +142,12 @@ TEST(StreamRange, Distance) {
 
 TEST(StreamRange, StreamError) {
   auto counter = 0;
-  auto reader = [&counter]() -> StreamReader<int>::result_type {
+  auto reader = [&counter]() -> internal::StreamReader<int>::result_type {
     if (counter++ < 2) return counter;
     return Status(StatusCode::kUnknown, "oops");
   };
 
-  StreamRange<int> sr = MakeStreamRange<int>(std::move(reader));
+  StreamRange<int> sr = internal::MakeStreamRange<int>(std::move(reader));
 
   auto it = sr.begin();
   EXPECT_NE(it, sr.end());
@@ -172,7 +171,6 @@ TEST(StreamRange, StreamError) {
 }
 
 }  // namespace
-}  // namespace internal
 }  // namespace GOOGLE_CLOUD_CPP_NS
 }  // namespace cloud
 }  // namespace google
